+++ crates/flashblocks-rpc/src/state.rs
@@ -122,7 +122,7 @@ mod tests {
     use crate::subscription::Metadata;
     use alloy_consensus::Receipt;
     use alloy_primitives::map::foldhash::HashMap;
-    use alloy_primitives::{Address, Bytes, B256, U256};
+    use alloy_primitives::{Address, Bloom, Bytes, B256, B64, U256};
     use alloy_rpc_types_engine::PayloadId;
     use op_alloy_consensus::OpDepositReceipt;
     use reth_optimism_chainspec::OpChainSpecBuilder;
@@ -265,6 +265,89 @@ mod tests {
             diff: delta2,
             metadata: metadata2,
         }
+    }
+
+    fn default_fb() -> Flashblock {
+        let block_info_tx = Bytes::from_str(
+            "0x7ef90104a06c0c775b6b492bab9d7e81abdf27f77cafb698551226455a82f559e0f93fea3794deaddeaddeaddeaddeaddeaddeaddeaddead00019442000000000000000000000000000000000000158080830f424080b8b0098999be000008dd00101c1200000000000000020000000068869d6300000000015f277f000000000000000000000000000000000000000000000000000000000d42ac290000000000000000000000000000000000000000000000000000000000000001abf52777e63959936b1bf633a2a643f0da38d63deffe49452fed1bf8a44975d50000000000000000000000005050f69a9786f081509234f1a7f4684b5e5b76c9000000000000000000000000")
+            .unwrap();
+
+        let block_info_txn_hash =
+            B256::from_str("0xba56c8b0deb460ff070f8fca8e2ee01e51a3db27841cc862fdd94cc1a47662b6")
+                .unwrap();
+
+        Flashblock {
+            payload_id: PayloadId(B64::random()),
+            index: 0,
+            base: Some(ExecutionPayloadBaseV1::default()),
+            diff: ExecutionPayloadFlashblockDeltaV1 {
+                state_root: B256::random(),
+                receipts_root: B256::random(),
+                logs_bloom: Bloom::random(),
+                gas_used: 1000,
+                block_hash: B256::random(),
+                transactions: vec![block_info_tx],
+                withdrawals: vec![],
+                withdrawals_root: B256::random(),
+            },
+            metadata: Metadata {
+                receipts: {
+                    let mut receipts = HashMap::default();
+                    receipts.insert(
+                        block_info_txn_hash,
+                        OpReceipt::Deposit(OpDepositReceipt {
+                            inner: Receipt {
+                                status: true.into(),
+                                cumulative_gas_used: 10000,
+                                logs: vec![],
+                            },
+                            deposit_nonce: Some(4012991u64),
+                            deposit_receipt_version: None,
+                        }),
+                    );
+                    receipts
+                },
+                new_account_balances: HashMap::default(),
+                block_number: 10,
+            },
+        }
+    }
+    #[test]
+    fn test_get_balance() {
+        let cs = Arc::new(OpChainSpecBuilder::base_mainnet().build());
+
+        let alice = Address::random();
+        let bob = Address::random();
+
+        let mut fb1 = default_fb();
+        fb1.metadata
+            .new_account_balances
+            .insert(alice, U256::from(1000));
+
+        let cache = PendingBlock::new_block(cs, fb1);
+        assert_eq!(
+            cache.get_balance(alice).expect("should be set"),
+            U256::from(1000)
+        );
+        assert!(cache.get_balance(bob).is_none());
+
+        let mut fb2 = default_fb();
+        fb2.metadata
+            .new_account_balances
+            .insert(alice, U256::from(2000));
+        fb2.metadata
+            .new_account_balances
+            .insert(bob, U256::from(1000));
+
+        let cache = PendingBlock::extend_block(&cache, fb2);
+        assert_eq!(
+            cache.get_balance(alice).expect("should be set"),
+            U256::from(2000)
+        );
+        assert_eq!(
+            cache.get_balance(bob).expect("should be set"),
+            U256::from(1000)
+        );
     }
 
     #[test]
